+++ packages/typespec-rust/src/codegen/unions.ts
@@ -8,6 +8,15 @@ import * as helpers from './helpers.js';
 import * as rust from '../codemodel/index.js';
 import {Context} from "./context.js";
 
+/** contains different types of content to emit */
+export interface Models {
+  /** types that are part of public surface area */
+  public?: helpers.Module;
+
+  /** trait impls for public types */
+  impls?: helpers.Module;
+}
+
 /**
  * returns the union enum types, or undefined if the
  * crate contains no union types.
@@ -16,12 +25,12 @@ import {Context} from "./context.js";
  * @param context the context for the provided crate
  * @returns the union content or undefined
  */
-export function emitUnions(crate: rust.Crate, context: Context): helpers.Module | undefined {
+export function emitUnions(crate: rust.Crate, context: Context): Models {
   if (crate.unions.length === 0) {
-    return undefined;
+    return {};
   }
 
-  const use = new Use('clients');
+  const use = new Use('modelsOther');
   const indent = new helpers.indentation();
 
   let body = '';
@@ -34,11 +43,8 @@ export function emitUnions(crate: rust.Crate, context: Context): helpers.Module 
     use.add('serde', 'Deserialize', 'Serialize');
     use.add('azure_core::fmt', 'SafeDebug');
     body += `#[derive(Deserialize, Serialize, SafeDebug, Clone)]\n`;
-    const tag = (rustUnion.discriminatorName !== '') ? `tag = "${rustUnion.discriminatorName}"` : '';
-    const content = (rustUnion.envelopeName !== '') ? `content = "${rustUnion.envelopeName}"` : '';
-    if (tag !== '' || content !== '') {
-      body += `#[serde(${[tag, content].filter(x => x !== '').join(', ')})]\n`;
-    }
+    const content = rustUnion.envelopeName ? `content = "${rustUnion.envelopeName}"` : '';
+    body += `#[serde(${[content, `tag = "${rustUnion.discriminant}"`].filter(x => x !== '').join(', ')})]\n`;
     body += `pub enum ${rustUnion.name} {\n`;
 
     for (let i = 0; i < rustUnion.members.length; ++i) {
@@ -51,18 +57,14 @@ export function emitUnions(crate: rust.Crate, context: Context): helpers.Module 
         body += `${indent.get()}#[doc = r#"${docs.substring(0, docs.length - 1)}"#]\n`;
       }
 
-      if (member.discriminatorValue !== member.name) {
-        body += `#[serde(rename = "${member.discriminatorValue}")]\n`;
+      if (member.discriminantValue !== member.type.name) {
+        body += `#[serde(rename = "${member.discriminantValue}")]\n`;
       }
-      body += `${indent.get()}${member.name}(${helpers.getTypeDeclaration(memberType)})`;
+      body += `${indent.get()}${member.type.name}(${helpers.getTypeDeclaration(memberType)})`;
       body += ',\n\n';
     }
 
     body += '}\n\n'; // end enum declaration
-  }
-
-  for (const rustUnion of crate.unions) {
-    body += context.getTryFromForRequestContent(rustUnion, use);
   }
 
   let content = helpers.contentPreamble();
@@ -70,7 +72,48 @@ export function emitUnions(crate: rust.Crate, context: Context): helpers.Module 
   content += body;
 
   return {
-    name: 'unions',
+    public: {
+      name: 'unions',
+      content: content,
+    },
+    impls: emitUnionImpls(crate, context),
+  };
+}
+
+/**
+ * returns any trait impls for union types.
+ * if no helpers are required, undefined is returned.
+ * 
+ * @param crate the crate for which to emit model impls
+ * @param context the context for the provided crate
+ * @returns the union impls content or undefined
+ */
+function emitUnionImpls(crate: rust.Crate, context: Context): helpers.Module | undefined {
+  const use = new Use('modelsOther');
+  const entries = new Array<string>();
+
+  for (const rustUnion of crate.unions) {
+    const forReq = context.getTryFromForRequestContent(rustUnion, use);
+
+    // helpers aren't required for all types, so only
+    // add a use statement for a type if it has a helper
+    if (forReq) {
+      use.addForType(rustUnion);
+      entries.push(forReq);
+    }
+  }
+
+  if (entries.length === 0) {
+    // no helpers
+    return undefined;
+  }
+
+  let content = helpers.contentPreamble();
+  content += use.text();
+  content += entries.sort().join('');
+
+  return {
+    name: 'unions_impl',
     content: content,
   };
 }