+++ packages/typespec-rust/src/codemodel/types.ts
@@ -18,7 +18,7 @@ export interface Docs {
 export type SdkType =  Arc | AsyncResponse | Box | ExternalType | ImplTrait | MarkerType | Option | Pager | Poller | RawResponse | RequestContent | Response | Result | Struct | TokenCredential | Unit;
 
 /** WireType defines types that go across the wire */
-export type WireType = Bytes | Decimal | EncodedBytes | Enum | EnumValue | Union | UnionMember | Etag | HashMap | JsonValue | Literal | Model | OffsetDateTime | RefBase | SafeInt | Scalar | Slice | StringSlice | StringType | Url | Vector;
+export type WireType = Bytes | Decimal | DiscriminatedUnion | EncodedBytes | Enum | EnumValue | Etag | HashMap | JsonValue | Literal | Model | OffsetDateTime | RefBase | SafeInt | Scalar | Slice | StringSlice | StringType | Url | Vector;
 
 /** Type defines a type within the Rust type system */
 export type Type = SdkType | WireType;
@@ -93,8 +93,8 @@ export interface Enum {
   /** any docs for the type */
   docs: Docs;
 
-  /** indicates if the enum and its values should be public */
-  pub: boolean;
+  /** indicates the visibility of the enum */
+  visibility: Visibility;
 
   /** one or more values for the enum */
   values: Array<EnumValue>;
@@ -123,47 +123,41 @@ export interface EnumValue {
   value: number | string;
 }
 
-/** Union is a Rust enum type with values of different types. */
-export interface Union {
-  kind: 'union';
+/** DiscriminatedUnion is a Rust tagged enum type */
+export interface DiscriminatedUnion {
+  kind: 'discriminatedUnion';
 
-  /** the name of the union type */
+  /** the name of the discriminated union */
   name: string;
 
   /** any docs for the type */
   docs: Docs;
 
-  /** indicates the visibility of the union */
+  /** indicates the visibility of the type */
   visibility: Visibility;
 
-  /** indicates if the union and its values should be public */
-  pub: boolean;
-
-  /** one or more members of the union */
-  members: Array<UnionMember>;
+  /** one or more members of the discriminated union */
+  members: Array<DiscriminatedUnionMember>;
 
   /** discriminator property name */
-  discriminatorName: string;
+  discriminant: string;
 
   /** data envelope property name */
-  envelopeName: string;
+  envelopeName?: string;
 }
 
-/** UnionMember is a union member for a specific Union */
-export interface UnionMember {
-  kind: 'unionMember';
-
-  /** the name of the union member */
-  name: string;
+/** DiscriminatedUnionMember is a tagged enum member for a specific DiscriminatedUnion */
+export interface DiscriminatedUnionMember {
+  kind: 'discriminatedUnionMember';
 
   /** any docs for the type */
   docs: Docs;
 
-  /** the type of the union member */
-  type: Type;
+  /** the type of the discriminated union member */
+  type: Model;
 
   /** discriminator property value */
-  discriminatorValue: string;
+  discriminantValue: string;
 }
 
 /** Etag is an azure_core::Etag */
@@ -304,7 +298,7 @@ export enum ModelFlags {
 }
 
 /** DateTimeEncoding is the wire format of the date/time */
-export type DateTimeEncoding = 'rfc3339' | 'rfc7231' | 'unix_time';
+export type DateTimeEncoding = 'rfc3339' | 'rfc3339-fixed-width' | 'rfc7231' | 'unix_time';
 
 /** OffsetDateTime is a Rust time::OffsetDateTime type */
 export interface OffsetDateTime extends External {
@@ -678,10 +672,10 @@ export class EncodedBytes implements EncodedBytes {
 }
 
 export class Enum implements Enum {
-  constructor(name: string, pub: boolean, extensible: boolean, type: EnumType) {
+  constructor(name: string, visibility: Visibility, extensible: boolean, type: EnumType) {
     this.kind = 'enum';
     this.name = name;
-    this.pub = pub;
+    this.visibility = visibility;
     this.values = new Array<EnumValue>();
     this.extensible = extensible;
     this.type = type;
@@ -758,23 +752,21 @@ export class MarkerType implements MarkerType {
   }
 }
 
-export class Union implements Union {
-  constructor(name: string, pub: boolean, discriminatorName: string, envelopeName: string) {
-    this.kind = 'union';
+export class DiscriminatedUnion implements DiscriminatedUnion {
+  constructor(name: string, visibility: Visibility, discriminant: string) {
+    this.kind = 'discriminatedUnion';
     this.name = name;
-    this.pub = pub;
-    this.members = new Array<UnionMember>();
-    this.discriminatorName = discriminatorName;
-    this.envelopeName = envelopeName;
+    this.visibility = visibility;
+    this.members = new Array<DiscriminatedUnionMember>();
+    this.discriminant = discriminant;
   }
 }
 
-export class UnionMember implements UnionMember {
-  constructor(name: string, type: Type, discriminatorValue: string) {
-    this.kind = 'unionMember';
-    this.name = name;
+export class DiscriminatedUnionMember implements DiscriminatedUnionMember {
+  constructor(type: Model, discriminantValue: string) {
+    this.kind = 'discriminatedUnionMember';
     this.type = type;
-    this.discriminatorValue = discriminatorValue;
+    this.discriminantValue = discriminantValue;
   }
 }